pyproject.toml
[project]
name = "py-eudr"
version = "0.1.0"
description = ""
authors = [
    {name = "Sergey Kovalskiy",email = "sergey@thorgate.eu"}
]
readme = "README.md"
requires-python = ">=3.8.6,<4.0"
dependencies = [
    "zeep (>=4.3.1,<=5.0.0)",
    "xmlschema (>=3.4.3,<=4.0.0)",
    "backports-zoneinfo[tzdata] (>=0.2.1,<0.3.0); python_version<'3.9'",
    "backports-strenum (>=1.2.8,<2.0.0); python_version<'3.11'"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
ruff = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
